LPFE-87: Moved basket.service to models
